reverted usb-ethernet setup to beagleboard scripts

+++ images/chroot_scripts/oresat-chroot.sh
@@ -34,18 +34,6 @@ echo "#USB Gadget Serial Port" >> /etc/securetty
 echo "ttyGS0" >> /etc/securetty
 
 ##############################################################################
-echo "setup usb ethernet"
-
-cat > "/etc/default/bb-boot" <<-__EOF__
-USB_NETWORK_DISABLED=yes
-USB_CONFIGURATION=enable
-__EOF__
-
-echo "g_ether" > /etc/modules-load.d/g_ether.conf
-HOST_ADDR=`dmesg | tr -s " " | grep "usb0: HOST MAC" | cut -d " " -f 6`
-echo "options g_ether host_addr=$HOST_ADDR" > /etc/modprobe.d/g_ether.conf
-
-##############################################################################
 echo "setup systemd-networkd"
 
 cat > "/etc/systemd/network/10-can.network" <<-__EOF__
@@ -56,9 +44,21 @@ Name=can1
 BitRate=1M
 __EOF__
 
-cat > "/etc/systemd/network/20-wired.network" <<-__EOF__
+cat > "/etc/systemd/network/20-wired-usb0.network" <<-__EOF__
 [Match]
 Name=usb0
+
+[Network]
+Address=192.168.7.2/24
+Gateway=192.168.7.1
+DNS=192.168.7.1
+DNS=8.8.8.8
+DNS=8.8.4.4
+__EOF__
+
+cat > "/etc/systemd/network/20-wired-usb1.network" <<-__EOF__
+[Match]
+Name=usb1
 
 [Network]
 Address=192.168.6.2/24
@@ -75,4 +75,6 @@ systemctl enable systemd-resolved
 ##############################################################################
 echo "remove internet packages required during build"
 
+if [ hostname != "dev" ]; then
 apt -y purge git git-man curl wget rsync
+fi
